+++ CHANGES.md
@@ -1,6 +1,6 @@
 0.12.0
 --------------
-- Add support for the rest of the required and recommended JWE algorithms (dir, A127KW, A256KW, ECDH-ES, ECDH-ES+A128KW)
+- Add support for the rest of the required and recommended JWE algorithms (dir, A127KW, A256KW, ECDH-ES, ECDH-ES+A128KW, ECDH-ES+A256KW)
 
 0.11.0
 --------------